% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{add_projflow_report}
\alias{add_projflow_report}
\title{Add a projflow report from RStudio}
\usage{
add_projflow_report()
}
\value{
Invisibly returns the created report path.
}
\description{
Add a projflow report from RStudio
}
\examples{
\dontrun{
add_projflow_report()
}
}
\author{
Thiago de Paula Oliveira
}
