% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{add_projflow_script}
\alias{add_projflow_script}
\title{Add a projflow script from RStudio}
\usage{
add_projflow_script()
}
\value{
Invisibly returns the created script path.
}
\description{
Add a projflow script from RStudio
}
\examples{
\dontrun{
add_projflow_script()
}
}
\author{
Thiago de Paula Oliveira
}
