% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{add_projflow_task}
\alias{add_projflow_task}
\title{Add a projflow task from RStudio}
\usage{
add_projflow_task()
}
\value{
Invisibly returns the created task identifier.
}
\description{
Add a projflow task from RStudio
}
\examples{
\dontrun{
add_projflow_task()
}
}
\author{
Thiago de Paula Oliveira
}
