% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{check_github_actions}
\alias{check_github_actions}
\title{Check GitHub Actions workflow files}
\usage{
check_github_actions(root = ".")
}
\arguments{
\item{root}{Existing project root whose GitHub Actions workflow files should
be inspected.}
}
\value{
Data frame describing known workflows.
}
\description{
Check GitHub Actions workflow files
}
\examples{
\dontrun{
check_github_actions()
}
}
\author{
Thiago de Paula Oliveira
}
