% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{close_project_risk}
\alias{close_project_risk}
\title{Close a project risk}
\usage{
close_project_risk(risk, root = ".")
}
\arguments{
\item{risk}{Risk id or title.}

\item{root}{Existing project root.}
}
\value{
Invisibly returns the updated risk id.
}
\description{
Close a project risk
}
\examples{
\dontrun{
close_project_risk("risk_0001")
}
}
\author{
Thiago de Paula Oliveira
}
