% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{complete_project_task}
\alias{complete_project_task}
\title{Complete a project task}
\usage{
complete_project_task(task, root = ".")
}
\arguments{
\item{task}{Task id or task title.}

\item{root}{Existing project root.}
}
\value{
Invisibly returns the updated task id.
}
\description{
Complete a project task
}
\examples{
\dontrun{
complete_project_task("task_0001")
}
}
\author{
Thiago de Paula Oliveira
}
