% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rproj_file.R
\name{create_rproj_file}
\alias{create_rproj_file}
\title{Create an RStudio project file}
\usage{
create_rproj_file(path, project_name, overwrite = FALSE)
}
\arguments{
\item{path}{Project root path.}

\item{project_name}{Project name.}

\item{overwrite}{Logical. Should an existing file be overwritten?}
}
\value{
A list describing the write outcome.
}
\description{
Create an RStudio project file
}
