% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{dashboard_status}
\alias{dashboard_status}
\title{Report background dashboard status}
\usage{
dashboard_status(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
A data frame with the recorded dashboard process status.
}
\description{
Report background dashboard status
}
\examples{
\dontrun{
dashboard_status()
}
}
\author{
Thiago de Paula Oliveira
}
