% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{find_project_root}
\alias{find_project_root}
\title{Find a project root directory}
\usage{
find_project_root(root = ".")
}
\arguments{
\item{root}{Starting path from which \code{projflow} should search upwards for a
project marker such as \code{project.yml}, \code{.projflow/project_registry.yml},
\code{.here}, or an \code{.Rproj} file.}
}
\value{
Absolute project root path.
}
\description{
Find a project root directory
}
\examples{
\dontrun{
projflow:::find_project_root(".")
}
}
\author{
Thiago de Paula Oliveira
}
