% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_git.R
\name{init_git_repo}
\alias{init_git_repo}
\title{Initialise a Git repository when available}
\usage{
init_git_repo(path, strict = FALSE)
}
\arguments{
\item{path}{Project root path where a Git repository should be initialised if
one does not already exist.}

\item{strict}{Logical scalar. If \code{TRUE}, Git initialisation failures raise an
error; otherwise a warning-style message is returned.}
}
\value{
\code{NULL} on success or a warning message when Git was not initialised.
}
\description{
Initialise a Git repository when available
}
\examples{
\dontrun{
tmp <- tempfile("projflow-git-init-")
dir.create(tmp)
projflow:::init_git_repo(tmp)
}
}
\author{
Thiago de Paula Oliveira
}
