% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_plan.R
\name{inspect_project_component}
\alias{inspect_project_component}
\title{Inspect a project component specification}
\usage{
inspect_project_component(component)
}
\arguments{
\item{component}{Character scalar. Component name to inspect.}
}
\value{
A structured component-specification object with class
\code{"projflow_component_spec"}.
}
\description{
Inspect a project component specification
}
\examples{
inspect_project_component("quality_control")
}
\author{
Thiago de Paula Oliveira
}
