% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_plan.R
\name{inspect_project_deliverable}
\alias{inspect_project_deliverable}
\title{Inspect a project deliverable specification}
\usage{
inspect_project_deliverable(deliverable)
}
\arguments{
\item{deliverable}{Character scalar. Deliverable name to inspect.}
}
\value{
A structured deliverable-specification object with class
\code{"projflow_deliverable_spec"}.
}
\description{
Inspect a project deliverable specification
}
\examples{
inspect_project_deliverable("client_report")
}
\author{
Thiago de Paula Oliveira
}
