% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_plan.R
\name{inspect_project_infrastructure}
\alias{inspect_project_infrastructure}
\title{Inspect project infrastructure support}
\usage{
inspect_project_infrastructure(infrastructure)
}
\arguments{
\item{infrastructure}{Character scalar. Infrastructure name to inspect.}
}
\value{
A structured infrastructure-specification object with class
\code{"projflow_infrastructure_spec"}.
}
\description{
Inspect project infrastructure support
}
\examples{
inspect_project_infrastructure("targets")
}
\author{
Thiago de Paula Oliveira
}
