% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_plan.R
\name{inspect_project_preset}
\alias{inspect_project_preset}
\title{Inspect a project preset}
\usage{
inspect_project_preset(preset)
}
\arguments{
\item{preset}{Character scalar. Preset name to inspect.}
}
\value{
A structured preset-specification object with class
\code{"projflow_preset_spec"}.
}
\description{
Inspect a project preset
}
\examples{
inspect_project_preset("client_report")
}
\author{
Thiago de Paula Oliveira
}
