% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_safety.R
\name{list_project_activity}
\alias{list_project_activity}
\title{List the project activity log}
\usage{
list_project_activity(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
Data frame of activity-log entries.
}
\description{
List the project activity log
}
\examples{
\dontrun{
list_project_activity()
}
}
\author{
Thiago de Paula Oliveira
}
