% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_safety.R
\name{list_project_backups}
\alias{list_project_backups}
\title{List available project backups}
\usage{
list_project_backups(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
Data frame of backup files.
}
\description{
List available project backups
}
\examples{
\dontrun{
list_project_backups()
}
}
\author{
Thiago de Paula Oliveira
}
