% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{list_project_reports}
\alias{list_project_reports}
\title{List registered project reports}
\usage{
list_project_reports(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
Data frame of registered reports.
}
\description{
List registered project reports
}
\examples{
\dontrun{
list_project_reports()
}
}
\author{
Thiago de Paula Oliveira
}
