% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{mark_project_milestone_done}
\alias{mark_project_milestone_done}
\title{Mark a project milestone as done}
\usage{
mark_project_milestone_done(milestone, root = ".")
}
\arguments{
\item{milestone}{Milestone id or title.}

\item{root}{Existing project root.}
}
\value{
Invisibly returns the updated milestone id.
}
\description{
Mark a project milestone as done
}
\examples{
\dontrun{
mark_project_milestone_done("milestone_0001")
}
}
\author{
Thiago de Paula Oliveira
}
