% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{new_project_object}
\alias{new_project_object}
\title{Create and register a project object}
\usage{
new_project_object(name, type, root = ".", overwrite = FALSE)
}
\arguments{
\item{name}{Object name.}

\item{type}{Object type.}

\item{root}{Project root.}

\item{overwrite}{Should an existing file be overwritten?}
}
\value{
Invisibly returns the created path.
}
\description{
Create and register a project object
}
