% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{open_dashboard}
\alias{open_dashboard}
\title{Open the projflow dashboard}
\usage{
open_dashboard(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}} when
\code{background = FALSE}.}
}
\value{
Invisibly returns the dashboard process handle when launched in the
background, or the app object when launched in the foreground.
}
\description{
Open the projflow dashboard
}
\examples{
\dontrun{
open_dashboard()
}
}
\author{
Thiago de Paula Oliveira
}
