% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{open_project_control_centre}
\alias{open_project_control_centre}
\title{Open the projflow Project Control Centre}
\usage{
open_project_control_centre(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Invisibly returns the launched \code{shiny.appobj}.
}
\description{
Open the projflow Project Control Centre
}
\examples{
\dontrun{
open_project_control_centre()
}
}
\author{
Thiago de Paula Oliveira
}
