% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{open_projflow_project_manager}
\alias{open_projflow_project_manager}
\title{Open the projflow Project Manager addin}
\usage{
open_projflow_project_manager()
}
\value{
Invisibly returns the launched dashboard.
}
\description{
Open the projflow Project Manager addin
}
\examples{
\dontrun{
open_projflow_project_manager()
}
}
\author{
Thiago de Paula Oliveira
}
