% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_repair.R
\name{project_check_items}
\alias{project_check_items}
\title{List machine-readable project check items}
\usage{
project_check_items(root = ".", deep = FALSE, render_reports = FALSE)
}
\arguments{
\item{root}{Existing project root.}

\item{deep}{Logical scalar passed to \code{check_project()}.}

\item{render_reports}{Logical scalar passed to \code{check_project()}.}
}
\value{
Data frame of project check items.
}
\description{
List machine-readable project check items
}
\examples{
\dontrun{
project_check_items()
}
}
\author{
Thiago de Paula Oliveira
}
