% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{project_data_root}
\alias{project_data_root}
\title{Read a configured external data root}
\usage{
project_data_root(name = "default", root = ".")
}
\arguments{
\item{name}{Data source name to retrieve from \code{.projflow/local.yml}.}

\item{root}{Existing project root.}
}
\value{
Absolute path.
}
\description{
Read a configured external data root
}
\examples{
\dontrun{
project_data_root()
project_data_root("reference")
}
}
\author{
Thiago de Paula Oliveira
}
