% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{project_decisions}
\alias{project_decisions}
\title{List project decisions}
\usage{
project_decisions(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
A data frame with one row per decision.
}
\description{
List project decisions
}
\examples{
\dontrun{
project_decisions()
}
}
\author{
Thiago de Paula Oliveira
}
