% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{project_milestones}
\alias{project_milestones}
\title{List project milestones}
\usage{
project_milestones(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
A data frame with one row per milestone.
}
\description{
List project milestones
}
\examples{
\dontrun{
project_milestones()
}
}
\author{
Thiago de Paula Oliveira
}
