% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_diagnostics.R
\name{project_network_data}
\alias{project_network_data}
\title{Construct project network data}
\usage{
project_network_data(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
A list with \code{nodes} and \code{edges} data frames.
}
\description{
Construct project network data
}
\examples{
\dontrun{
network <- project_network_data()
names(network)
}
}
\author{
Thiago de Paula Oliveira
}
