% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{project_paths}
\alias{project_paths}
\title{Return standard project paths}
\usage{
project_paths(root = ".")
}
\arguments{
\item{root}{Project root. Defaults to the current project.}
}
\value{
A named list of absolute project paths.
}
\description{
Return standard project paths
}
