% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{project_risks}
\alias{project_risks}
\title{List project risks}
\usage{
project_risks(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
A data frame with one row per risk.
}
\description{
List project risks
}
\examples{
\dontrun{
project_risks()
}
}
\author{
Thiago de Paula Oliveira
}
