% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{project_status_report}
\alias{project_status_report}
\title{Create a project status report}
\usage{
project_status_report(root = ".", output = c("markdown", "html", "data"))
}
\arguments{
\item{root}{Existing project root.}

\item{output}{Output format to return.}
}
\value{
Either a markdown string, HTML string, or a structured list.
}
\description{
Create a project status report
}
\examples{
\dontrun{
project_status_report(output = "data")
}
}
\author{
Thiago de Paula Oliveira
}
