% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{project_tasks}
\alias{project_tasks}
\title{List project tasks}
\usage{
project_tasks(root = ".")
}
\arguments{
\item{root}{Existing project root.}
}
\value{
A data frame with one row per task.
}
\description{
List project tasks
}
\examples{
\dontrun{
project_tasks()
}
}
\author{
Thiago de Paula Oliveira
}
