% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{remove_project_milestone}
\alias{remove_project_milestone}
\title{Remove a project milestone}
\usage{
remove_project_milestone(milestone, root = ".")
}
\arguments{
\item{milestone}{Milestone id or title.}

\item{root}{Existing project root.}
}
\value{
Invisibly returns the removed milestone id.
}
\description{
Remove a project milestone
}
\examples{
\dontrun{
remove_project_milestone("milestone_0001")
}
}
\author{
Thiago de Paula Oliveira
}
