% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{remove_project_risk}
\alias{remove_project_risk}
\title{Remove a project risk}
\usage{
remove_project_risk(risk, root = ".")
}
\arguments{
\item{risk}{Risk id or title.}

\item{root}{Existing project root.}
}
\value{
Invisibly returns the removed risk id.
}
\description{
Remove a project risk
}
\examples{
\dontrun{
remove_project_risk("risk_0001")
}
}
\author{
Thiago de Paula Oliveira
}
