% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{remove_project_task}
\alias{remove_project_task}
\title{Remove a project task}
\usage{
remove_project_task(task, root = ".")
}
\arguments{
\item{task}{Task id or task title.}

\item{root}{Existing project root.}
}
\value{
Invisibly returns the removed task id.
}
\description{
Remove a project task
}
\examples{
\dontrun{
remove_project_task("task_0001")
}
}
\author{
Thiago de Paula Oliveira
}
