% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{resolve_project_path}
\alias{resolve_project_path}
\title{Validate a project path}
\usage{
resolve_project_path(path)
}
\arguments{
\item{path}{Target project path.}
}
\value{
A normalized project path.
}
\description{
Validate a project path
}
