% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{run_projflow_check}
\alias{run_projflow_check}
\title{Run the projflow project check addin}
\usage{
run_projflow_check()
}
\value{
Invisibly returns the printed project-check result.
}
\description{
Run the projflow project check addin
}
\examples{
\dontrun{
run_projflow_check()
}
}
\author{
Thiago de Paula Oliveira
}
