% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{save_external_project_object}
\alias{save_external_project_object}
\title{Save a project object to external storage}
\usage{
save_external_project_object(
  object,
  name,
  type,
  source = "default",
  root = "."
)
}
\arguments{
\item{object}{Object to save.}

\item{name}{Object name.}

\item{type}{Object type.}

\item{source}{External data source name.}

\item{root}{Project root.}
}
\value{
Invisibly returns the saved path.
}
\description{
Save a project object to external storage
}
