% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_code_loading.R
\name{scaffold_code_loading}
\alias{scaffold_code_loading}
\title{Create code loading scaffold files}
\usage{
scaffold_code_loading(
  path,
  project_name,
  code_loading = c("package", "box", "source"),
  template_data = list(),
  overwrite = FALSE
)
}
\arguments{
\item{path}{Project root path.}

\item{project_name}{Project name.}

\item{code_loading}{Code loading strategy.}

\item{template_data}{Template data used for rendering.}

\item{overwrite}{Logical. Should existing files be overwritten?}
}
\value{
A list with created and skipped file paths.
}
\description{
Create code loading scaffold files
}
