% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_quarto.R
\name{scaffold_quarto}
\alias{scaffold_quarto}
\title{Create Quarto scaffold files}
\usage{
scaffold_quarto(path, project_name, template_data = list(), overwrite = FALSE)
}
\arguments{
\item{path}{Project root path.}

\item{project_name}{Project name.}

\item{template_data}{Template data used for rendering.}

\item{overwrite}{Logical. Should existing files be overwritten?}
}
\value{
A list with created and skipped file paths.
}
\description{
Create Quarto scaffold files
}
