% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_targets.R
\name{scaffold_targets}
\alias{scaffold_targets}
\title{Create targets scaffold files}
\usage{
scaffold_targets(path, project_name, template_data = list(), overwrite = FALSE)
}
\arguments{
\item{path}{Project root path.}

\item{project_name}{Project name.}

\item{template_data}{Template data used for rendering.}

\item{overwrite}{Logical. Should existing files be overwritten?}
}
\value{
A list with created and skipped file paths.
}
\description{
Create targets scaffold files
}
