% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_package_names}
\alias{validate_package_names}
\title{Validate package names}
\usage{
validate_package_names(packages, allow_null = TRUE)
}
\arguments{
\item{packages}{Package vector.}

\item{allow_null}{Logical. Whether \code{NULL} is allowed.}
}
\value{
Cleaned package names.
}
\description{
Validate package names
}
