% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_project_name}
\alias{validate_project_name}
\title{Validate a project name}
\usage{
validate_project_name(project_name)
}
\arguments{
\item{project_name}{Project name supplied by the user.}
}
\value{
A validated project name.
}
\description{
Validate a project name
}
