% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_project_path}
\alias{validate_project_path}
\title{Validate whether a project path can be used}
\usage{
validate_project_path(path, overwrite = FALSE)
}
\arguments{
\item{path}{Target project path.}

\item{overwrite}{Logical. Should existing files be overwritten?}
}
\value{
Invisibly returns the validated path.
}
\description{
Validate whether a project path can be used
}
