% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_templates.R
\name{write_template_file}
\alias{write_template_file}
\title{Write a template file}
\usage{
write_template_file(path, content, overwrite = FALSE)
}
\arguments{
\item{path}{Destination file path.}

\item{content}{File content.}

\item{overwrite}{Logical. Should an existing file be overwritten?}
}
\value{
A list describing the write outcome.
}
\description{
Write a template file
}
