% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_api.R
\name{build_project}
\alias{build_project}
\title{Build a project}
\usage{
build_project(root = ".", render_reports = TRUE, run_apps = FALSE)
}
\arguments{
\item{root}{Project root.}

\item{render_reports}{Should registered reports be rendered?}

\item{run_apps}{Should the app be launched after building?}
}
\value{
Structured build results.
}
\description{
Build a project
}
