% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{check_project}
\alias{check_project}
\title{Check whether a project is ready to run}
\usage{
check_project(root = ".")
}
\arguments{
\item{root}{Project root. Defaults to the current project.}
}
\value{
An object of class \code{"project_status"}.
}
\description{
Check whether a project is ready to run
}
