% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_git_files.R
\name{create_git_files}
\alias{create_git_files}
\title{Create Git support files}
\usage{
create_git_files(path, overwrite = FALSE)
}
\arguments{
\item{path}{Project root path.}

\item{overwrite}{Logical. Should existing files be overwritten?}
}
\value{
A list with created and skipped file paths.
}
\description{
Create Git support files
}
