% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{find_project_root}
\alias{find_project_root}
\title{Find a project root directory}
\usage{
find_project_root(root = ".")
}
\arguments{
\item{root}{Starting path.}
}
\value{
Absolute project root path.
}
\description{
Find a project root directory
}
