% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_git.R
\name{init_git_repo}
\alias{init_git_repo}
\title{Initialise a Git repository when available}
\usage{
init_git_repo(path, strict = FALSE)
}
\arguments{
\item{path}{Project root path.}

\item{strict}{Logical. Should failures error instead of warning?}
}
\value{
\code{NULL} on success or a warning message when Git was not initialised.
}
\description{
Initialise a Git repository when available
}
