% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_renv.R
\name{init_renv_project}
\alias{init_renv_project}
\title{Initialise renv when available}
\usage{
init_renv_project(
  path,
  packages = character(),
  snapshot = TRUE,
  strict = FALSE
)
}
\arguments{
\item{path}{Project root path.}

\item{packages}{Character vector of packages to install into the project
library after initialising renv.}

\item{snapshot}{Logical. Should \code{renv::snapshot()} be run?}

\item{strict}{Logical. Should failures error instead of warning?}
}
\value{
\code{NULL} on success or a warning message when renv was not initialised.
}
\description{
Initialise renv when available
}
