% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{list_project_data_sources}
\alias{list_project_data_sources}
\title{List configured external data sources}
\usage{
list_project_data_sources(root = ".")
}
\arguments{
\item{root}{Existing project root whose local data-source configuration should
be listed.}
}
\value{
Data frame of configured sources.
}
\description{
List configured external data sources
}
\examples{
\dontrun{
list_project_data_sources()
}
}
\author{
Thiago de Paula Oliveira
}
