% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{list_project_objects}
\alias{list_project_objects}
\title{List project objects}
\usage{
list_project_objects(root = ".")
}
\arguments{
\item{root}{Project root. Defaults to the current project.}
}
\value{
A data frame describing registered project objects.
}
\description{
List project objects
}
