% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{list_project_outputs}
\alias{list_project_outputs}
\title{List registered project outputs}
\usage{
list_project_outputs(root = ".")
}
\arguments{
\item{root}{Existing project root whose registered outputs should be listed.}
}
\value{
Data frame of outputs and their existence status.
}
\description{
List registered project outputs
}
\examples{
\dontrun{
list_project_outputs()
}
}
\author{
Thiago de Paula Oliveira
}
