% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_api.R
\name{new_app}
\alias{new_app}
\title{Create a new app or dashboard}
\usage{
new_app(
  name = "app",
  type = c("shiny", "quarto_dashboard"),
  root = ".",
  open = interactive()
)
}
\arguments{
\item{name}{App or dashboard name.}

\item{type}{App type.}

\item{root}{Project root.}

\item{open}{Included for API compatibility.}
}
\value{
Invisibly returns the created app or dashboard path.
}
\description{
Create a new app or dashboard
}
