% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_api.R
\name{new_component}
\alias{new_component}
\title{Add a component to an existing project}
\usage{
new_component(component, root = ".", open = interactive())
}
\arguments{
\item{component}{Component name.}

\item{root}{Project root.}

\item{open}{Included for API compatibility.}
}
\value{
A scaffold result object.
}
\description{
Add a component to an existing project
}
