% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_api.R
\name{new_report}
\alias{new_report}
\title{Create a new report}
\usage{
new_report(
  name = "main_report",
  type = c("html_report", "client_report", "scientific_report"),
  root = ".",
  open = interactive()
)
}
\arguments{
\item{name}{Report name.}

\item{type}{Report type.}

\item{root}{Project root.}

\item{open}{Included for API compatibility.}
}
\value{
Invisibly returns the created report path.
}
\description{
Create a new report
}
