% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level_api.R
\name{new_script}
\alias{new_script}
\title{Create a new project script}
\usage{
new_script(
  name,
  type = "analysis",
  root = ".",
  order = NULL,
  open = interactive()
)
}
\arguments{
\item{name}{Script name.}

\item{type}{Script type.}

\item{root}{Project root.}

\item{order}{Optional execution order.}

\item{open}{Included for API compatibility.}
}
\value{
Invisibly returns the created script path.
}
\description{
Create a new project script
}
