% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{print.project_status}
\alias{print.project_status}
\title{Print a project status summary}
\usage{
\method{print}{project_status}(x, ...)
}
\arguments{
\item{x}{A \code{"project_status"} object, usually created by \code{project_status()}.}

\item{...}{Additional arguments accepted for S3 compatibility but ignored by
this method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a project status summary
}
\examples{
\dontrun{
x <- project_status()
print(x)
}
}
\author{
Thiago de Paula Oliveira
}
