% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{project_status}
\alias{project_status}
\title{Check project status}
\usage{
project_status(root = ".")
}
\arguments{
\item{root}{Project root. Defaults to the current project.}
}
\value{
An object of class \code{"project_status"}.
}
\description{
Check project status
}
