% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_runtime.R
\name{use_github_actions}
\alias{use_github_actions}
\title{Add a GitHub Actions workflow}
\usage{
use_github_actions(root = ".", workflow = c("check-project", "render-reports"))
}
\arguments{
\item{root}{Project root.}

\item{workflow}{Workflow name.}
}
\value{
Invisibly returns the created workflow path.
}
\description{
Add a GitHub Actions workflow
}
