% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_character_vector}
\alias{validate_character_vector}
\title{Validate a character vector}
\usage{
validate_character_vector(x, arg, allow_null = FALSE)
}
\arguments{
\item{x}{Object to validate.}

\item{arg}{Argument name.}

\item{allow_null}{Logical. Whether \code{NULL} is allowed.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Validate a character vector
}
