% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_choice}
\alias{validate_choice}
\title{Validate a single choice}
\usage{
validate_choice(x, choices, arg)
}
\arguments{
\item{x}{Choice to validate.}

\item{choices}{Allowed choices.}

\item{arg}{Argument name.}
}
\value{
The validated choice.
}
\description{
Validate a single choice
}
