% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_logical_scalar}
\alias{validate_logical_scalar}
\title{Validate a logical scalar}
\usage{
validate_logical_scalar(x, arg)
}
\arguments{
\item{x}{Object to validate.}

\item{arg}{Argument name.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Validate a logical scalar
}
