% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_package_names}
\alias{validate_package_names}
\title{Validate package names}
\usage{
validate_package_names(packages, allow_null = TRUE)
}
\arguments{
\item{packages}{Character vector of package names to validate and clean.}

\item{allow_null}{Logical scalar indicating whether \code{NULL} is accepted.}
}
\value{
Cleaned package names.
}
\description{
Validate package names
}
\examples{
projflow:::validate_package_names(c("fs", "yaml", "fs"))
}
\author{
Thiago de Paula Oliveira
}
