% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_project_name}
\alias{validate_project_name}
\title{Validate a project name}
\usage{
validate_project_name(project_name)
}
\arguments{
\item{project_name}{Project name supplied by the user, usually derived from a
folder name or explicit project title field.}
}
\value{
A validated project name.
}
\description{
Validate a project name
}
\examples{
projflow:::validate_project_name("My analysis project")
}
\author{
Thiago de Paula Oliveira
}
